requirements.txt
argh==0.26.2
astroid==2.0.4
autopep8==1.4
certifi==2017.7.27.1
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathtools==0.1.2
pycodestyle==2.4.0
pycurl==7.43.0
pylint==2.1.1
PyVimeo==1.0.9
PyYAML==3.13
requests==2.19.1
six==1.10.0
tinydb==3.11.1
tuspy==0.2.1
typed-ast==1.1.0
urllib3==1.23
watchdog==0.9.0
wrapt==1.10.11
